pyproject.toml
[tool.poetry]
name = "studyme-backend"
version = "0.1.0"
description = ""
authors = ["Iurii Nefedov <y.nefedov.work@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.76.0"
pydantic = "^1.9.0"
SQLAlchemy = "^1.4.36"
alembic = "^1.7.7"
SQLAlchemy-Utils = "^0.38.2"
starlette-exporter = "^0.12.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.20.0"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.25.0"
greenlet = "^1.1.2"
passlib = "^1.7.4"
python-jose = "^3.3.0"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
